+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 - Slider: Fit thumbs within the bar area not to cause overflow
 - Slider: Fixed horizontal scrolling on mount in safari
 - Actionable: Switched to :focus-visible
+- Badge: Relaxed the types to be able to pass dismissAriaLabel while onDismiss is undefined
+- Modal: Added full-screen position